# static/swagger-ui/swagger/openapi.yaml
swagger: "2.0"
info:
  title: JeeekAPI
  description: エンジニアのための自己実現SNSアプリ
  contact:
    name: tonouchi510
    email: tonouchi27@gmail.com
    url: https://github.com/tonouchi510/Jeeek/issues
  version: "0.1"
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /v1/activity/qiita/{user_id}:
    get:
      tags:
      - Activity
      summary: Fetch qiita article by qiita-user-id Activity
      description: 指定したユーザのQiitaの記事投稿を取得する
      operationId: Activity#Fetch qiita article by qiita-user-id
      parameters:
      - name: user_id
        in: path
        description: user id of qiita
        required: true
        type: string
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/ActivityFetchQiitaArticleByQiitaUserIDUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
definitions:
  ActivityFetchQiitaArticleByQiitaUserIDUnauthorizedResponseBody:
    title: ActivityFetchQiitaArticleByQiitaUserIDUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Culpa vel deleniti ut sed.
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: Secures endpoint by requiring a valid JWT token retrieved via the
      firebase.
    name: Authorization
    in: header
...